pyproject.toml
[project]
name = "extractor-energy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "mlx-vlm>=0.3.9",
    "pillow>=12.1.0",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
    "torchvision>=0.15.0",
    "pymupdf>=1.26.7",
    "ultralytics>=8.4.6",
    "paddleocr>=2.7.0",
]
